+++ src/ScottPlot5/ScottPlot5/MultiplotSharedAxisManager.cs
@@ -1,6 +1,6 @@
 namespace ScottPlot;
 
-public class MultiplotSharedAxisManager
+public class MultiplotSharedAxisManager : IMultiplotPreRenderAction
 {
     // TODO: improve support for plots with non-standard axis limits
     private readonly List<Plot> PlotsWithSharedX = [];
@@ -81,5 +81,10 @@ public class MultiplotSharedAxisManager
             }
         }
         return null;
+    }
+
+    public void Invoke()
+    {
+        UpdateSharedPlotAxisLimits();
     }
 }